1. update application.yml 2. move *buildustomizer to initializr-extension module
