+++ initializr-extension/pom.xml
@@ -20,6 +20,10 @@
     </dependency>
     <dependency>
       <groupId>io.spring.initializr</groupId>
+      <artifactId>initializr-generator-spring</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>io.spring.initializr</groupId>
       <artifactId>initializr-metadata</artifactId>
     </dependency>
     <dependency>
@@ -65,11 +69,6 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.spring.initializr</groupId>
-      <artifactId>initializr-generator-spring</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
